+++ tests/window/window.cpp
@@ -10,7 +10,7 @@ public:
         Window::Style::Resizable
     ) { }
 
-    bool OnUpdate(const float elapsed_time) override {
+    bool OnUpdate(const int64_t elapsed_time) override {
         return is_open();
     }
